+++ src/Pages/Cadastro/Cadastro.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+function Cadastro() {
+    return <h1>Cadastro</h1>;
+}
+
+export default Cadastro;
