+++ src/Pages/Login/Login.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function Login() {
+    return <h1>Login</h1>
+}
