+++ src/routes.js
@@ -1,0 +1,18 @@
+import React from "react";
+import { BrowserRouter, Route, Routes } from "react-router-dom";
+import Cadastro from "./Pages/Cadastro";
+import Home from "./Pages/Home";
+import Login from "./Pages/Login";
+
+function Router() {
+    return (
+    <BrowserRouter>
+        <Routes>
+            <Route path="/cadastro" element={<Cadastro/>}/>
+            <Route path="/home" element={<Home/>}/>
+            <Route path="/login" element={<Login/>}/>
+        </Routes>
+    </BrowserRouter>);
+}
+
+export default Router;
